fix: backend

+++ techPoints/techPoints/src/main/kotlin/techForAll/techPoints/dtos/CursoAlunosDto.kt
@@ -2,5 +2,5 @@ package techForAll.techPoints.dtos
 
 class CursoAlunosDto(
     val nomeCurso: String,
-    val quantidadeAlunos: Long
+    val quantidadeAlunos: Int
 )
